This PR updates the presentation of incompatible fields in the same family, and the presentation of custom fields, as detailed in <https://github.com/elastic/security-team/issues/6046>
Indices created by older versions of integrations or Beats may have mappings flagged as incompatible with the latest version of ECS, because their types don't exactly match, but their mappings are in the same _type family_.

The `Field data types` documentation <https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-types.html> introduces the concept of type families:

> Field types are grouped by _family_. Types in the same family have exactly the same search behavior but may have different space usage or performance characteristics.

In this PR, the Data Quality dashboard:

- Still flags a field as incompatible when the `ECS mapping type (expected)` doesn't exactly match the `Index mapping type (actual)`
- Applies a different visual treatment to incompatible fields in the same family (in the table of results, and in markdown)
- Has revised content in the `Incompatible fields` tab's callout (and exported markdown), to include the concept of families

It's perfectly OK to add custom fields to an index. This PR updates the presentation of the `Custom fields` tab callout to remove potentially confusing descriptions.

The official `ECS Reference Overview` documentation <https://www.elastic.co/guide/en/ecs/current/ecs-reference.html> states:

> ECS is a permissive schema. If your events have additional data that cannot be mapped to ECS, you can simply add them to your events, using custom field names.

In this PR, the Data Quality dashboard:

- Has revised content in the `Custom fields` tab's callout (and exported markdown), to remove potentially confusing descriptions, and convey the permissive nature of ECS.

The following screenshot illustrates the OLD behavior of the `Incompatable fields` tab:

![exact_matching](https://user-images.githubusercontent.com/4459398/219443483-03b891ae-aad0-4098-b88e-229725e65604.png)

In the screenshot of the OLD behavior above:

- When the `ECS mapping type (expected)` doesn't exactly match the `Index mapping type (actual)`, the field is counted as incompatible
- All results in the expanded index are styled with red ("eui danger") styles

The following new screenshot is marked up to show things that are (still) the SAME in the new behavior:

![things_that_stayed_the_same_20230217](https://user-images.githubusercontent.com/4459398/219804049-c18bcf8e-0b9a-4b9f-aed8-268b417e4aa8.png)

The screenshot of the NEW behavior above illustrates what remains the SAME:

- When the `ECS mapping type (expected)` doesn't exactly match the `Index mapping type (actual)`, the field is **still** flagged as incompatible
- The total counts of `Incompatable fields` in the pattern summary, table, and tabs are still the same
- The `Incompatible fields` callout has the same `Fields are incompatible with ECS when...` description, and items with an ❌
- The same fields in the OLD behavior that were flagged as incompatible are still being flagged

The following new screenshot is marked up to show things that are DIFFERENT:

![things_that_are_different_20230217](https://user-images.githubusercontent.com/4459398/219805550-193074d2-155c-4bae-822c-ae41f9571a39.png)

The screenshot of the NEW behavior above illustrates what's DIFFERENT:

- The title of the `Incompatible fields` callout has been updated to include a count of fields with mappings in the same family, e.g. `12 incompatible fields, 11 fields with mappings in the same family`
- The callout as been updated to include new content, which incorporates families
- The results in the expanded index are styled with yellow ("eui warning") styles, when the index type is in the same family
- The results in the expanded index also have a new `same family` badge, when the index type is in the same family

The following new screenshot illustrates that fields in the `Index mapping type (actual)` column still have the same red / "eui danger" styling, (and no badge), when they are not in the same family as the `ECS mapping type (expected)`:

![no_mappings_in_the_same_family_2023_0217](https://user-images.githubusercontent.com/4459398/219808700-40ea3b77-ce41-477c-a326-b4e90d8365ee.png)

This following screenshot is a side-by-side comparison between the Data Quality dashboard and the markdown created by the `Add to new case` and `Copy to clipboard` actions, to illustrate that it includes the changes made to the dashboard itself:

![incompatible_fields_markdown_updated_20230217](https://user-images.githubusercontent.com/4459398/219811354-f64515a0-01cd-48ce-99b0-d62a757006b4.png)

The following screenshot illustrates the OLD content of the `Custom fields` tab's callout:

![custom_fields_before](https://user-images.githubusercontent.com/4459398/219501783-297ec50c-627b-4c85-8aad-78ee46497143.png)

The following screenshot illustrates the NEW content of the `Custom fields` tab's callout:

![custom_fields_after](https://user-images.githubusercontent.com/4459398/219503486-f7938e72-1570-4b5c-8fc6-3b5fa3e141cd.png)

This following screenshot is a side-by-side comparison between the Data Quality dashboard and the markdown created by the  `Copy to clipboard` action, to illustrate that it includes the changes made to the dashboard itself:

![custom_fields_markdown_updated_20230217](https://user-images.githubusercontent.com/4459398/219815766-40743c80-c51f-4498-b87a-b80d5646926c.png)